pyproject.toml
[project]
name = "triptych-detector-trim"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "pikamonvvs", email = "pikamonvvs@naver.com"},
]
dependencies = ["ffmpeg-python>=0.2.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
